split up vulcan tests for per-commit actions workflow
